# Optimize optimizer eliminate stage