… tools/optimizer/CMakeLists.txt
… tools/optimizer/parser.h
… tools/optimizer/simple_ast.cpp
… tools/optimizer/simple_ast.h