@aidanhs 2015-09-02
All core asm2 tests succeed.