@kripken 2015-09-03
The engine isn't open, even as a .bc. But I'll test locally tomorrow with each patch separately, to at least try to narrow down the issue.

I'm hoping though that this isn't specific to unity. Maybe poppler or bananabread (I can provide a bc for that) with `-s LINKABLE=1` (to disable dead code elimination and leave a lot more stuff) might also show something.