@aidanhs 2015-09-03
I suspect you'll find it's the final commit, 31e528a5017. I should've really gone through the benchmarks before, will try and do so tomorrow.