@aidanhs 2015-09-07
I've added a couple of new commits after managing to grab the unity .bc file from a unity install and I'm mostly happy with where this is up to.

I did benchmarking of the optimizer run directly against an arbitrary unity js chunk (14M), the sqlite (14M) and the lambdas testcase (22M). Each measurement is for *all* O3 optimisation passes, best of 5 runs, user time as reported by the `time` bash builtin:

Case | Before branch | After branch
----|-------|-------
Unity | 27.431s | 26.247s
Sqlite | 1m20.520s | 1m18.458s
Lambdas | 1m19.287s | 52.067s

It's worth taking a look at "Avoid array bounds checking for traverse* functions" - it's theoretically less safe because Ref[] does bounds checks, but I feel that these functions are small enough to be verifiable.