@kripken 2015-09-08
I still see a slowdown here, on unity. When I set `EMCC_CORES=1` to reduce noise, I get `50.88 50.99` for the js opts stage in `EMCC_DEBUG=1` before this patch, and `51.12 51.20` after this patch. It's around a 0.5% slowdown, so not large, but still in the wrong direction.

How are you measuring your times?