@aidanhs 2015-09-09
Summary of changes:

 - at an explicit reminder (for people like me) to enable release mode when cmaking the optimizer
 - use a name->vec lookup for dependencies rather than name->set as vecs are faster
 - new commit to clean up ref construction

Now I've fixed my issues and tweaked the dep lookup, I get from about 54.1s knocked down to 53.3s for the unity js opts phase on a single core. `other.test_js_opts` passes.