@kripken 2015-09-09
I still see qualitatively different numbers here, no change or a tiny slowdown on unity, tiny slowdown on poppler, no change on bullet.

For poppler, for example, I run `EMCC_CORES=1 EMCC_DEBUG=1 ./emcc -O2 poppler.bc  -s LINKABLE=1` and I get `7.62 7.60 7.61` before, `7.67 7.68 7.67` after. Very small slowdown, but consistent.

Maybe it's just a matter of a different cpu / different local compiler and stdlib, that accounts for us seeing different things. But if so then it suggests the changes here are going to vary across machines.

If we can't figure this out, it might make sense to still merge this or parts of this, as the refactorings are nice. But perhaps we should only take the refactorings for clarity, and not for speed?