@kripken 2015-09-09
Another thought, there might be some bigger things to optimize in that pass. If I recall correctly, around the comment "`Look for statements, including while-switch pattern`" we have the main loop, which traverses the whole tree and calls `scan`. And `scan` does its own traversal on the entire tree shown to it. I don't remember all the details, but I wonder if we couldn't stop those traversals in some cases, and avoid a big chunk of `O(N^2)` work. But, this is tricky stuff.