@aidanhs 2015-09-17
I've pushed a bunch of changes. I opted not to go for the harder idea, and instead pursued the low-hanging fruit. There are some speedups from

 - Reserving space in vectors on creation, rather than relying on pushing to reallocate ("Reserve space on vector creation", "Also reserve space when creation nodes during runtime").
 - Allocate vectors in an arena, like Values ("Allocate vecs in arena").
 - Improving memory locality in traverse* by sticking values in locals ("Caching values in locals is faster than hitting memory", "Cache array size, use pointer arithmetic to bypass some indirection").

Unfortunately, the final item makes the traverse* functions really ugly.

However, the rest of the commits are generally just refactoring. You may wish to examine the following with suspicion:

 - "Don't pass optional blocks as arguments" - removes optional parameters, in exchange for needing a function to transfer block contents from one node to another. I prefer this way because it makes it very clear that the functions are always going to create a block.
 - "Don't perform unnecessary handling of VAR nodes" :warning: - the commit message provides justification, but was it deliberate that processing would continue in release mode (i.e. when the assert was removed)? I figured that requiring valid asm was a small price to pay for the code simplification.

The rest of the tweaks should be easily verifiable.

I see a ~10% speedup in the optimisation of unity (54.5s to 48.5s with `EMCC_CORES=1 EMCC_DEBUG=1 emcc -O2 UnityNative.bc -s LINKABLE=1`), sqlite (6.45s to 5.87s with `asm2.test_sqlite`) and poppler (4.97 to 4.48 with `asm2.test_poppler`).

I've run all asm2 tests (+`other.test_{js,native}_optimizer`) with assertions enabled on the optimizer and they seem fine (anything that fails also fails on incoming for me).