@kripken 2015-09-18
Confirmed, I see an 8% speedup. Nice! Reviewing now.