@kripken 2015-09-18
Please add any changes after this point as followup commits, so I don't need to re-read the ones I am going through now.