@kripken 2015-09-18
Merged, thanks!